+++ src/main/java/com/seeme/service/api/MicrodustOpenApi.java
@@ -1,6 +1,7 @@
 package com.seeme.service.api;
 
 import com.google.gson.Gson;
+import com.seeme.config.ApiConfig;
 import com.seeme.domain.microdust.*;
 import com.seeme.util.JSONParsingUtil;
 import com.seeme.util.MicrodustUtil;
