+++ src/main/java/com/seeme/util/ErrorMessageUtil.java
@@ -1,6 +1,6 @@
 package com.seeme.util;
 
-public class ErrorMessage {
+public class ErrorMessageUtil {
 
 	public static final String API_PROVISION_ERROR = "API_PROVISION_ERROR";
 	public static final String JSON_PARSING_ERROR = "JSON_PARSING_ERROR";
